// src/api/generated/schema.ts
export interface components {
    schemas: {
        SageKahootApplicationSessionsAnswerAcceptedResult: {
            /** Format: guid */
            sessionId?: string;
            /** Format: guid */
            questionId?: string;
            /** Format: int64 */
            responseTimeMs?: number;
        };
        MicrosoftAspNetCoreMvcProblemDetails: {
            type?: string | null;
            title?: string | null;
            /** Format: int32 */
            status?: number | null;
            detail?: string | null;
            instance?: string | null;
        } & {
            [key: string]: unknown;
        };
// ...
        FastEndpointsProblemDetails: {
            /** @default https://www.rfc-editor.org/rfc/rfc7231#section-6.5.1 */
            type: string;
            /** @default One or more validation errors occurred. */
            title: string;
            /**
             * Format: int32
             * @default 400
             */
            status: number;
            /** @default /api/route */
            instance: string;
            /** @default 0HMPNHL0JHL76:00000001 */
            traceId: string;
            /** @description the details of the error */
            detail?: string | null;
            errors?: components["schemas"]["FastEndpointsProblemDetails_Error"][];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}